third_party/WebKit/LayoutTests/external/wpt/audio-output/setSinkId.https-expected.txt

This is a testharness.js-based test.
PASS setSinkId on default audio output should always work
PASS setSinkId fails with NotFoundError on made up deviceid
PASS List media devices
FAIL Correctly reacts to setting known deviceid as sinkid 0 assert_equals: On known devices, the only possible failure of setSinkId is a securityerror expected "SecurityError" but got "NotFoundError"
FAIL Correctly reacts to setting known deviceid as sinkid 1 assert_equals: On known devices, the only possible failure of setSinkId is a securityerror expected "SecurityError" but got "NotFoundError"
FAIL Correctly reacts to setting known deviceid as sinkid 2 assert_equals: On known devices, the only possible failure of setSinkId is a securityerror expected "SecurityError" but got "NotFoundError"
Harness: the test ran to completion.
